# docs/public/api/openapi-3.0.json
{
    "openapi": "3.0.0",
    "info": {
        "description": "Track, Manage, and Organize your Things.",
        "title": "Homebox API",
        "contact": {
            "name": "Homebox Team",
            "url": "https://discord.homebox.software"
        },
        "version": "1.0"
    },
    "paths": {
        "/v1/actions/create-missing-thumbnails": {
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Creates thumbnails for items that are missing them",
                "tags": [
                    "Actions"
                ],
                "summary": "Create Missing Thumbnails",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/v1.ActionAmountResult"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/v1/actions/ensure-asset-ids": {
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Ensures all items in the database have an asset ID",
                "tags": [
                    "Actions"
                ],
                "summary": "Ensure Asset IDs",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/v1.ActionAmountResult"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/v1/actions/ensure-import-refs": {
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Ensures all items in the database have an import ref",
                "tags": [
                    "Actions"
                ],
                "summary": "Ensures Import Refs",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/v1.ActionAmountResult"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/v1/actions/set-primary-photos": {
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Sets the first photo of each item as the primary photo",
                "tags": [
                    "Actions"
                ],
                "summary": "Set Primary Photos",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/v1.ActionAmountResult"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/v1/actions/wipe-inventory": {
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Rejected: offboard each item or place with a disposition instead.",
                "tags": [
                    "Actions"
                ],
                "summary": "Wipe Inventory",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/v1.WipeInventoryOptions"
                            }
                        }
                    },
                    "description": "Wipe options"
                },
                "responses": {
                    "409": {
                        "description": "Conflict",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/validate.ErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/v1/actions/zero-item-time-fields": {
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Resets all item date fields to the beginning of the day",
                "tags": [
                    "Actions"
                ],
                "summary": "Zero Out Time Fields",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/v1.ActionAmountResult"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/v1/assets/{id}": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "tags": [
                    "Items"
                ],
                "summary": "Get Item by Asset ID",
                "parameters": [
                    {
                        "description": "Asset ID",
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/repo.PaginationResult-repo_EntitySummary"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/v1/entities": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "tags": [
                    "Entities"
                ],
                "summary": "Query All Entities",
                "parameters": [
                    {
                        "description": "search string",
                        "name": "q",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "page number",
                        "name": "page",
                        "in": "query",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "description": "items per page",
                        "name": "pageSize",
                        "in": "query",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "description": "tags Ids",
                        "name": "tags",
                        "in": "query",
                        "explode": true,
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        }
                    },
                    {
                        "description": "parent Ids",
                        "name": "parentIds",
                        "in": "query",
                        "explode": true,
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/repo.EntityListResult"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "tags": [
                    "Entities"
                ],
                "summary": "Create Entity",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/repo.EntityCreate"
                            }
                        }
                    },
                    "description": "Entity Data",
                    "required": true
                },
                "responses": {
                    "201": {
                        "description": "Created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/repo.EntityOut"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/v1/entities/export": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "tags": [
                    "Entities"
                ],
                "summary": "Export Entities",
                "responses": {
                    "200": {
                        "description": "text/csv",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/v1/entities/fields": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "tags": [
                    "Entities"
                ],
                "summary": "Get All Custom Field Names",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/v1/entities/fields/values": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "tags": [
                    "Entities"
                ],
                "summary": "Get All Custom Field Values",
                "parameters": [
                    {
                        "description": "Field name",
                        "name": "field",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/v1/entities/import": {
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "tags": [
                    "Entities"
                ],
                "summary": "Import Entities",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "csv": {
                                        "description": "CSV file to upload",
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "required": [
                                    "csv"
                                ]
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "204": {
                        "description": "No Content"
                    }
                }
            }
        },
        "/v1/entities/tree": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "tags": [
                    "Entities"
                ],
                "summary": "Get Locations Tree",
                "parameters": [
                    {
                        "description": "include items in response tree",
                        "name": "withItems",
                        "in": "query",
                        "schema": {
                            "type": "boolean"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/repo.TreeItem"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/v1/entities/{id}/path": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "tags": [
                    "Entities"
                ],
                "summary": "Get the full path of an entity",
                "parameters": [
                    {
                        "description": "Entity ID",
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/repo.EntityPath"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/v1/users/logout": {
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "tags": [
                    "Authentication"
                ],
                "summary": "User Logout",
                "responses": {
                    "204": {
                        "description": "No Content"
                    }
                }
            }
        },
...
    },
    "servers": [
        {
            "url": "/api"
        }
    ],
    "components": {
        "securitySchemes": {
            "Bearer": {
                "description": "\"Type 'Bearer TOKEN' to correctly set the API Key\"",
                "type": "apiKey",
                "name": "Authorization",
                "in": "header"
            }
        },
        "schemas": {
...
            "v1.ActionAmountResult": {
                "type": "object",
                "properties": {
                    "completed": {
                        "type": "integer"
                    }
                }
            },
...
            "v1.WipeInventoryOptions": {
                "type": "object",
                "properties": {
                    "wipeLocations": {
                        "type": "boolean"
                    },
                    "wipeMaintenance": {
                        "type": "boolean"
                    },
                    "wipeTags": {
                        "type": "boolean"
                    }
                }
            },
...
            "validate.ErrorResponse": {
                "type": "object",
                "properties": {
                    "error": {
                        "type": "string"
                    },
                    "fields": {
                        "type": "string"
                    }
                }
            }
        }
    }
}